pyproject.toml
[tool.poetry]
name = "weight_tracker"
version = "0.1.0"
description = "A simple way to track body weight over time"
authors = ["EricEhmke <eric@ericehmke.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
marshmallow = "^3.14.1"
pandas = "^1.4.1"
Flask-SQLAlchemy = "^2.5.1"
Flask-Login = "^0.5.0"
flask-marshmallow = "^0.14.0"
python-dotenv = "^0.19.2"
marshmallow-sqlalchemy = "^0.27.0"
PyJWT = "^2.3.0"
pymongo = "^4.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
